refactor: move db queries to a single location
